[rewrite] Implement `q = N` case in blind rotation (#17)
* fix: remove .vscode

* feat: implement rgsw_by_rgsw and extend allocate_scratch

* feat: implement q = N case in lmkcdey

+++ math/src/poly/automorphism.rs
@@ -74,7 +74,7 @@ mod test {
         for log_n in 0..10 {
             let n = 1 << log_n;
             let indices = (0..n as i64).collect_vec();
-            for k in powers_mod(5, 2 * n as u64).take(n / 2) {
+            for k in powers_mod(5, 2 * n).take(n / 2) {
                 let auto_map = AutomorphismMap::new(n, k as _);
                 assert_eq!(
                     auto_map.apply(&indices, |i| -i).collect_vec(),
